Cambio de int a Single en Tasa de Cambio para arreglar el Controller

+++ API_MercaditoTEC/API_MercaditoTEC/Models/TasaCambio.cs
@@ -8,7 +8,7 @@ namespace API_MercaditoTEC.Models
         [Key]
         public int idTasaCambio { get; set; }
         [Required]
-        public int monto { get; set; } //Cambiar int por Single cuando se cambie en la base de datos
+        public Single monto { get; set; } //Cambiar int por Single cuando se cambie en la base de datos
         [Required]
         public DateTime fechaPublicacion { get; set; }
     }
